apps/api/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Auth
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
requests==2.31.0

# Storage
boto3==1.34.18

# PDF Generation
reportlab==4.0.9
# weasyprint==60.2  # Alternative HTML to PDF

# API Documentation
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
openpyxl==3.1.2
pandas==2.1.4
